requirements.txt
anthropic>=0.40.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0

# Dev dependencies
pytest>=7.0
pytest-asyncio>=0.23.0
